chore: move outsideExecution class to utils
